refactor: consolidate zip entry name utilities and enhance public API tests
