+++ src/viewport/api/public.py
@@ -1,8 +1,5 @@
 import contextlib
-import re
-import unicodedata
 from asyncio import run as asyncio_run
-from pathlib import Path
 from uuid import UUID
 
 import zipstream
@@ -11,7 +8,6 @@ from fastapi.responses import StreamingResponse
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from viewport.api.photo import CONTENT_TYPE_MAP
 from viewport.dependencies import get_s3_client as get_async_s3_client
 from viewport.logger import logger
 from viewport.models.db import get_db
@@ -23,6 +19,7 @@ from viewport.s3_utils import get_s3_client, get_s3_settings
 from viewport.schemas.gallery import GalleryPhotoSortBy, SortOrder
 from viewport.schemas.public import PublicCover, PublicGalleryResponse, PublicPhoto
 from viewport.sharelink_utils import is_sharelink_expired
+from viewport.zip_utils import build_zip_fallback_name, make_unique_zip_entry_name, sanitize_zip_entry_name
 
 router = APIRouter(prefix="/s", tags=["public"])
 PUBLIC_CACHE_CONTROL_HEADERS = {
@@ -31,139 +28,10 @@ PUBLIC_CACHE_CONTROL_HEADERS = {
     "Expires": "0",
 }
 
-_WINDOWS_RESERVED_NAMES = {
-    "con",
-    "prn",
-    "aux",
-    "nul",
-    "com1",
-    "com2",
-    "com3",
-    "com4",
-    "com5",
-    "com6",
-    "com7",
-    "com8",
-    "com9",
-    "lpt1",
-    "lpt2",
-    "lpt3",
-    "lpt4",
-    "lpt5",
-    "lpt6",
-    "lpt7",
-    "lpt8",
-    "lpt9",
-}
-_SEPARATOR_LIKE_CHARS = {"/", "\\", "\u2215", "\u2044"}
-_FORBIDDEN_ZIP_CHARS_RE = re.compile(r'[<>:"|?*\x00-\x1F]')
-_WHITESPACE_RE = re.compile(r"\s+")
-_WINDOWS_DRIVE_PREFIX_RE = re.compile(r"^[A-Za-z]:")
-_MAX_ZIP_ENTRY_NAME_BYTES = 255
-_EXTENSION_ONLY_TOKENS = {ext.lstrip(".") for ext in CONTENT_TYPE_MAP}
-
 
 def _build_content_disposition(filename: str, disposition_type: str = "inline") -> str:
     safe_filename = filename.replace("\\", "\\\\").replace('"', '\\"')
     return f'{disposition_type}; filename="{safe_filename}"'
-
-
-def _split_name_and_ext(filename: str) -> tuple[str, str]:
-    path = Path(filename)
-    suffix = path.suffix if path.suffix else ""
-    stem = path.stem if path.stem else "file"
-    return stem, suffix
-
-
-def _truncate_utf8(value: str, max_bytes: int) -> str:
-    if max_bytes <= 0:
-        return ""
-    encoded = value.encode("utf-8")
-    if len(encoded) <= max_bytes:
-        return value
-    truncated = encoded[:max_bytes]
-    return truncated.decode("utf-8", errors="ignore")
-
-
-def _truncate_preserving_extension(filename: str, max_bytes: int = _MAX_ZIP_ENTRY_NAME_BYTES) -> str:
-    if len(filename.encode("utf-8")) <= max_bytes:
-        return filename
-
-    stem, suffix = _split_name_and_ext(filename)
-    suffix_bytes = len(suffix.encode("utf-8"))
-
-    if suffix_bytes >= max_bytes:
-        return _truncate_utf8(stem, max_bytes)
-
-    stem_max_bytes = max_bytes - suffix_bytes
-    truncated_stem = _truncate_utf8(stem, stem_max_bytes).rstrip(" .")
-    if not truncated_stem:
-        truncated_stem = "file"
-    return f"{truncated_stem}{suffix}"
-
-
-def _sanitize_zip_entry_name(filename: str, fallback: str) -> str:
-    normalized = unicodedata.normalize("NFKC", filename or "")
-    cleaned = "".join(ch for ch in normalized if unicodedata.category(ch)[0] != "C").strip()
-    if not cleaned:
-        return fallback
-
-    if any(separator in cleaned for separator in _SEPARATOR_LIKE_CHARS):
-        return fallback
-
-    if _WINDOWS_DRIVE_PREFIX_RE.match(cleaned):
-        return fallback
-
-    sanitized = _FORBIDDEN_ZIP_CHARS_RE.sub("_", cleaned)
-    sanitized = _WHITESPACE_RE.sub(" ", sanitized).strip(" .")
-
-    if not sanitized or sanitized in {".", ".."}:
-        return fallback
-
-    if "." not in sanitized and sanitized.casefold() in _EXTENSION_ONLY_TOKENS:
-        return fallback
-
-    stem, _ = _split_name_and_ext(sanitized)
-    if stem.casefold() in _WINDOWS_RESERVED_NAMES:
-        return fallback
-
-    sanitized = _truncate_preserving_extension(sanitized)
-    if not sanitized or sanitized in {".", ".."}:
-        return fallback
-
-    return sanitized
-
-
-def _build_zip_fallback_name(filename: str, object_key: str, fallback_stem: str) -> str:
-    normalized = unicodedata.normalize("NFKC", filename or "")
-    leaf = normalized.replace("\\", "/").rsplit("/", 1)[-1]
-    extension = Path(leaf).suffix.lower()
-    if extension in CONTENT_TYPE_MAP:
-        return f"{fallback_stem}{extension}"
-
-    object_key_extension = Path(object_key).suffix.lower()
-    if object_key_extension in CONTENT_TYPE_MAP:
-        return f"{fallback_stem}{object_key_extension}"
-
-    return f"{fallback_stem}.jpg"
-
-
-def _make_unique_zip_entry_name(filename: str, used_names: set[str]) -> str:
-    candidate = filename
-    stem, suffix = _split_name_and_ext(filename)
-    counter = 1
-
-    while candidate.casefold() in used_names:
-        suffix_part = f" ({counter})"
-        stem_budget = _MAX_ZIP_ENTRY_NAME_BYTES - len(suffix.encode("utf-8"))
-        candidate_stem = _truncate_utf8(f"{stem}{suffix_part}", stem_budget).rstrip(" .")
-        if not candidate_stem:
-            candidate_stem = "file"
-        candidate = f"{candidate_stem}{suffix}"
-        counter += 1
-
-    used_names.add(candidate.casefold())
-    return candidate
 
 
 def _resolve_public_sorting(gallery: Gallery) -> tuple[GalleryPhotoSortBy, SortOrder]:
@@ -340,9 +208,9 @@ def download_all_photos_zip(
 
     for photo in photos:
         key = photo.object_key
-        fallback = _build_zip_fallback_name(photo.display_name, object_key=key, fallback_stem=f"photo-{photo.id}")
-        filename = _sanitize_zip_entry_name(photo.display_name, fallback=fallback)
-        filename = _make_unique_zip_entry_name(filename, used_names)
+        fallback = build_zip_fallback_name(photo.display_name, object_key=key, fallback_stem=f"photo-{photo.id}")
+        filename = sanitize_zip_entry_name(photo.display_name, fallback=fallback)
+        filename = make_unique_zip_entry_name(filename, used_names)
 
         def file_generator(object_key: str = key):
             client = get_s3_client()
