+++ src/viewport/filename_utils.py
@@ -1,4 +1,5 @@
 import re
+from pathlib import Path
 
 
 def sanitize_filename(filename: str) -> str:
@@ -18,3 +19,38 @@ def sanitize_filename(filename: str) -> str:
         filename = "file"
 
     return filename
+
+
+def split_name_and_ext(filename: str) -> tuple[str, str]:
+    path = Path(filename)
+    suffix = path.suffix if path.suffix else ""
+    stem = path.stem if path.stem else "file"
+    return stem, suffix
+
+
+def truncate_utf8(value: str, max_bytes: int) -> str:
+    if max_bytes <= 0:
+        return ""
+
+    encoded = value.encode("utf-8")
+    if len(encoded) <= max_bytes:
+        return value
+
+    return encoded[:max_bytes].decode("utf-8", errors="ignore")
+
+
+def truncate_preserving_extension(filename: str, max_bytes: int = 255) -> str:
+    if len(filename.encode("utf-8")) <= max_bytes:
+        return filename
+
+    stem, suffix = split_name_and_ext(filename)
+    suffix_bytes = len(suffix.encode("utf-8"))
+
+    if suffix_bytes >= max_bytes:
+        return truncate_utf8(stem, max_bytes)
+
+    stem_max_bytes = max_bytes - suffix_bytes
+    truncated_stem = truncate_utf8(stem, stem_max_bytes).rstrip(" .")
+    if not truncated_stem:
+        truncated_stem = "file"
+    return f"{truncated_stem}{suffix}"
